feat: teste

+++ src/main/java/com/hackaton/postech/useCase/implementation/ClientService.java
@@ -1,13 +1,19 @@
 package com.hackaton.postech.useCase.implementation;
 
 import com.hackaton.postech.application.exceptions.NotFoundException;
-import com.hackaton.postech.application.mapper.ClientMapper;
+import com.hackaton.postech.domain.dto.request.AddressRequestDTO;
 import com.hackaton.postech.domain.dto.request.ClientRequestDTO;
 import com.hackaton.postech.domain.dto.response.ClientResponseDTO;
+import com.hackaton.postech.domain.model.Address;
 import com.hackaton.postech.domain.model.Client;
+import com.hackaton.postech.domain.repository.AddressRepository;
 import com.hackaton.postech.domain.repository.ClientRepository;
 import com.hackaton.postech.useCase.contract.IClientService;
 import jakarta.persistence.EntityNotFoundException;
+import org.hibernate.boot.model.internal.CreateKeySecondPass;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.transaction.annotation.Transactional;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -18,53 +24,63 @@ import java.util.List;
 public class ClientService implements IClientService {
 
   private final ClientRepository repository;
-  private final ClientMapper mapper;
+  private final AddressRepository addressRepository;
 
-  public List<ClientResponseDTO> findAll() {
-    return repository.findAll()
-        .stream()
-        .map(mapper::convertToClientResponse)
-        .toList();
+  @Transactional(readOnly = true)
+  public Page<ClientResponseDTO> findAll(PageRequest pageRequest) {
+    var clients = repository.findAll(pageRequest);
+    return clients.map(ClientResponseDTO::of);
+
   }
-
+  @Transactional(readOnly = true)
   public ClientResponseDTO findById(Long clientId) {
-    return  mapper.convertToClientResponse(findByIdClient(clientId));
+
+    return repository.findById(clientId)
+        .stream()
+        .map(ClientResponseDTO::of)
+        .findFirst()
+        .orElseThrow(() -> new NotFoundException("Cliente não encontrado"));
   }
 
+  @Transactional
   public ClientResponseDTO save(ClientRequestDTO clientRequestDTO) {
-    var clientSaved = repository.save(mapper.convertToClient(clientRequestDTO));
 
-    return mapper.convertToClientResponse(clientSaved);
+    var addressSaved = addressRepository.save(Address.ofSave(clientRequestDTO.getAddress()));
+    var clientSaved = repository.save(Client.ofSave(clientRequestDTO, AddressRequestDTO.of(addressSaved)));
+
+    return ClientResponseDTO.of(clientSaved);
   }
 
+  @Transactional
   public ClientResponseDTO update(Long clientId, ClientRequestDTO clientRequestDTO) {
 
     try {
-      var client = repository.getReferenceById(clientId);
-      var clientUpdated = mapper.convertToClient(clientRequestDTO);
-      clientUpdated.setClientId(client.getClientId());
 
-      client = repository.save(clientUpdated);
-      return mapper.convertToClientResponse(client);
+      var clientUpdate = repository.getReferenceById(clientId);
+      var addressUpdate = Address.ofUpdate(clientRequestDTO.getAddress());
+      addressUpdate.setAddressId(clientUpdate.getClientId());
+      clientRequestDTO.setClientId(clientUpdate.getClientId());
+      ClientRequestDTO.mapperEntity(clientRequestDTO, clientUpdate, addressUpdate);
+
+      addressUpdate = addressRepository.save(addressUpdate);
+      clientUpdate = repository.save(clientUpdate);
+      return ClientResponseDTO.of(clientUpdate);
     } catch (EntityNotFoundException exception) {
       throw new NotFoundException("Cliente não encontrado, id: " + clientId);
     }
   }
 
+  @Transactional
   public void deleteById(Long clientId) {
 
-    try {
-      repository.findById(clientId);
+    if(repository.existsById(clientId)) {
       repository.deleteById(clientId);
-    } catch (EntityNotFoundException exception) {
-      throw new NotFoundException("Cliente não encontrado, id: " + clientId);
+    } else {
+      throw new NotFoundException("Recurso não encontrado");
     }
   }
 
-  private Client findByIdClient(Long clientId) {
-    return repository.findById(clientId)
-        .orElseThrow(() -> new NotFoundException("Cliente não encontrado na base de dados."));
-  }
+
